+++ flake.nix
@@ -6,7 +6,7 @@
     nixpkgs.url =
       "github:NixOS/nixpkgs/ed014c27f4d0ca772fb57d3b8985b772b0503bbd";
     hspkgs.url =
-      "github:podenv/hspkgs/24d2028871584f71313ac06e23ef143db61aea34";
+      "github:podenv/hspkgs/cd711c5967c1313b6f91d1d40c7d68bfd561cfbe";
   };
 
   outputs = { self, nixpkgs, hspkgs }: